Added validation and missing routes

+++ index.js
@@ -40,37 +40,49 @@ db.once('open', function() {
   /*
    * Handle Login
    */
+  router.get('/login', (req, res) => {
+    return res.send(400).send(Response.badRequest);
+  });
 
   router.post('/login', (req, res) => {
     const username = req.body.username;
     const password = req.body.password;
 
-    User.findOne({
-      username: username
-    }, (err, data) => {
-      if(data && password === data.password) {
-
-        const token = jwt.sign({
-          exp: Math.floor(Date.now() / 1000) + 3600,
-          username: username,
-        }, "JWT_SECRET");
-
-        return res.status(200).send({
-          "token": token,
+    if(!username || !password) {
+      return res.status(400).send(Response.badRequest);
+    } else {
+      User
+        .findOne({ username: username })
+        .catch( err => {
+          console.error(err);
+          return res.status(500).send(Response.serverError);
+        })
+        .then(data => {
+          if (data && password === data.password) {
+            return jwt.sign({
+              exp: Math.floor(Date.now() / 1000) + 3600,
+              username: username,
+            }, "JWT_SECRET");
+          } else {
+            return null
+          }
+        })
+        .then( token => {
+          if(token) {
+            return res.status(200).send({
+              "token": token,
+            });
+          } else {
+            res.status(401).send(Response.unauthorized);
+          }
         });
-
-      } else {
-        return res.status(401).send(Response.unauthorized);
-
-      }
-    });
+    }
   });
-  router.get('/login', (req, res) => {
-    return res.send(400).send(Response.badRequest);
-  });
+
   router.put('/login', (req, res) => {
     return res.send(400).send(Response.badRequest);
   });
+
   router.delete('/login', (req, res) => {
     return res.send(400).send(Response.badRequest);
   });
@@ -80,10 +92,10 @@ db.once('open', function() {
    */
 
   router.get('/contacts', JWTMiddleware, (req, res) => {
-    const sort = req.query.sort || 'desc';
+    const order = req.query.order || 'asc';
     const filter = req.query.filter || null;
 
-    const sortingParam = sort === 'asc' ? -1 : 1;
+    const sortingParam = order === 'asc' ? -1 : 1;
 
     let query = {};
 
@@ -108,6 +120,28 @@ db.once('open', function() {
       });
   });
 
+  router.get('/contacts/:id', JWTMiddleware, (req, res) => {
+    const contactId = req.params.id;
+
+    if(contactId && mongoose.Types.ObjectId.isValid(contactId)) {
+      Contact
+        .findOne({_id: contactId})
+        .catch( err => {
+          console.error(err);
+          return res.status(500).send(Response.serverError);
+        })
+        .then( data => {
+          if(data) {
+            return res.status(200).send(data);
+          } else {
+            return res.status(404).send(Response.notFound);
+          }
+        });
+    } else {
+      return res.status(404).send(Response.notFound);
+    }
+  });
+
   /*
    * Handle contact create
    */
@@ -117,31 +151,35 @@ db.once('open', function() {
     const lastName = req.body.lastName;
     const phoneNumber = req.body.phoneNumber;
 
-    Contact.findOne({firstName, lastName, phoneNumber})
-      .catch(err => {
-        console.error(err);
-        return res.status(500).send(Response.serverError);
-      })
-      .then(data => {
-        if(!data) {
-          const newContact = new Contact({
-            firstName,
-            lastName,
-            phoneNumber,
-          });
+    if(!firstName || !lastName || !phoneNumber) {
+      return res.status(400).send(Response.badRequest);
+    } else {
+      Contact.findOne({firstName, lastName, phoneNumber})
+        .catch(err => {
+          console.error(err);
+          return res.status(500).send(Response.serverError);
+        })
+        .then(data => {
+          if(!data) {
+            const newContact = new Contact({
+              firstName,
+              lastName,
+              phoneNumber,
+            });
 
-          return newContact.save();
-        } else {
-          return null;
-        }
-      })
-      .then((person) => {
-        if(!person) {
-          return res.status(400).send(Response.alreadyExists);
-        } else {
-          return res.status(201).send(person);
-        }
-      });
+            return newContact.save();
+          } else {
+            return null;
+          }
+        })
+        .then((person) => {
+          if(!person) {
+            return res.status(400).send(Response.alreadyExists);
+          } else {
+            return res.status(201).send(person);
+          }
+        });
+    }
   });
 
   /*
@@ -153,7 +191,7 @@ db.once('open', function() {
     const lastName = req.body.lastName;
     const phoneNumber = req.body.phoneNumber;
 
-    if(id) {
+    if(id && mongoose.Types.ObjectId.isValid(id)) {
       Contact.findByIdAndUpdate(id, {
         firstName,
         lastName,
@@ -179,13 +217,15 @@ db.once('open', function() {
       const updateQueue = entries.map( entry => {
         const entryId = entry._id;
 
-        return Contact.updateOne({ _id: entryId}, { ...entry })
-          .then( raw => raw)
-          .then( data => {
-            if(data.nModified === 1) {
-              updatedEntries.push(entry);
-            }
-          });
+        if(entryId && mongoose.Types.ObjectId.isValid(entryId)) {
+          return Contact.updateOne({ _id: entryId}, { ...entry })
+            .then( raw => raw)
+            .then( data => {
+              if(data.nModified === 1) {
+                updatedEntries.push(entry);
+              }
+            });
+        }
       });
 
 
@@ -206,9 +246,13 @@ db.once('open', function() {
   /*
    * Handle contact delete
    */
+  router.delete('/contacts', JWTMiddleware, (req, res) => {
+    return res.status(400).send(Response.badRequest);
+  });
+
   router.delete('/contacts/:id', JWTMiddleware, (req, res) => {
     const id = req.params.id;
-    if(id) {
+    if(id && mongoose.Types.ObjectId.isValid(id)) {
       Contact.findOneAndDelete({_id: id})
         .catch( err => {
           console.error(err);
